+++ src/frontend/src/common/queries/entities.ts
@@ -6,4 +6,5 @@ export enum ENTITIES {
   LINEAGES_INFO = "LINEAGES_INFO",
   GROUP_INFO = "GROUP_INFO",
   GROUP_MEMBER_INFO = "GROUP_MEMBER_INFO",
+  GROUP_INVITATION_INFO = "GROUP_INVITATION_INFO",
 }
