+++ src/frontend/src/common/routes.ts
@@ -6,7 +6,6 @@ export enum ROUTES {
   DATA = "/data",
   GISAID = "https://www.gisaid.org/",
   GITHUB = "https://github.com/chanzuckerberg/czgenepi/",
-  INVITATIONS = "/members/invitations",
   NEXTSTRAIN = "https://nextstrain.org/",
   TERMS = "/terms",
   PANGOLIN = "https://pangolin.cog-uk.io/",
@@ -26,4 +25,5 @@ export enum ROUTES {
   GROUP = "/group",
   GROUP_DETAILS = "/group/details",
   GROUP_MEMBERS = "/group/members",
+  GROUP_INVITATIONS = "/group/members/invitations",
 }
