+++ src/frontend/src/views/GroupMembersPage/index.tsx
@@ -33,7 +33,7 @@ const isValidPrimaryTab = (token?: string) => {
   return token === PrimaryTabType.MEMBERS || token === PrimaryTabType.DETAILS;
 };
 
-const GroupMembersPage = ({ pathTokens }: Props): JSX.Element => {
+const GroupMembersPage = ({ pathTokens }: Props): JSX.Element | null => {
   const [primaryQueryParam, secondaryQueryParam] = pathTokens ?? [];
   const initialPrimaryTab = (
     isValidPrimaryTab(primaryQueryParam)
@@ -47,32 +47,18 @@ const GroupMembersPage = ({ pathTokens }: Props): JSX.Element => {
   const { data: userInfo } = useUserInfo();
   const groupId = getGroupIdFromUser(userInfo);
   const { data: members = [] } = useGroupMembersInfo(groupId);
-  const { data: groupInfo = {} } = useGroupInfo(groupId);
-  const { address, location, name, prefix } = groupInfo;
+  const { data: groupInfo } = useGroupInfo(groupId);
 
   useEffect(() => {
     router.push(`${ROUTES.GROUP}/${tabValue}`, undefined, { shallow: true });
   }, [tabValue]);
 
-  // TODO (mlila): api calls
-  const invites = [
-    {
-      email: "erica@fake.com",
-      status: "pending",
-      dateSent: "2022-05-24",
-      role: "Member",
-    },
-    {
-      email: "frank@fake.com",
-      status: "expired",
-      dateSent: "2022-03-24",
-      role: "Member",
-    },
-  ];
+  if (!groupInfo) return null;
+
+  const { address, location, name, prefix } = groupInfo;
 
   // sort group members by name before display
   members.sort((a, b) => (a.name > b.name ? 1 : -1));
-  invites.sort((a, b) => (a.dateSent > b.dateSent ? 1 : -1));
 
   const displayLocation = stringifyGisaidLocation(location);
